+++ apps/functions-subscription/src/adapters/azure/functions/process-activations-changes.ts
@@ -1,18 +1,23 @@
 import * as t from 'io-ts';
-import * as O from 'fp-ts/lib/Option';
 import * as TE from 'fp-ts/lib/TaskEither';
-import * as RA from 'fp-ts/lib/ReadonlyArray';
 import { pipe } from 'fp-ts/lib/function';
 import { InvocationContext } from '@azure/functions';
 import { ActivationRequestCodec } from '../../../domain/activation-request';
 import { ActivationJobCodec } from '../../../domain/activation-job';
 import { SystemEnv } from '../../../system-env';
+import { Config } from '../../../config';
 
-export const makeActivationJobConsumerHandler =
-  (
-    env: Pick<SystemEnv, 'processActivationJob'>,
-    maxConcurrencyThreshold: number,
-  ) =>
+export const makeActivationsChangesHandler =
+  ({
+    env,
+    config,
+  }: {
+    readonly env: Pick<
+      SystemEnv,
+      'processActivationJob' | 'processActivationRequest'
+    >;
+    readonly config: Pick<Config, 'activations'>;
+  }) =>
   (
     documents: unknown,
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -25,14 +30,19 @@ export const makeActivationJobConsumerHandler =
           .array(t.union([ActivationRequestCodec, ActivationJobCodec]))
           .decode(documents),
       ),
-      // Keep only job documents
-      TE.map(
-        RA.filterMap((doc) => (doc.type === 'job' ? O.some(doc) : O.none)),
-      ),
       TE.flatMap(
-        TE.traverseArray((job) =>
-          env.processActivationJob(job, maxConcurrencyThreshold),
-        ),
+        TE.traverseArray((document) => {
+          if (document.type === 'job')
+            return env.processActivationJob(
+              document,
+              config.activations.maxFetchSize,
+            );
+          else
+            return pipe(
+              env.processActivationRequest(document),
+              TE.map(() => 'success' as const),
+            );
+        }),
       ),
       TE.getOrElse((error) => {
         // if an error occurs, the retry policy will be applied if it is defined