+++ apps/functions-subscription/src/use-cases/__tests__/process-subscription-request.test.ts
@@ -6,26 +6,32 @@ import {
   aSubscriptionRequest,
   aSubscription,
   aSubscriptionHistory,
+  anActivationRequest,
 } from '../../domain/__tests__/data';
 import { makeTestEnv } from '../../domain/__tests__/mocks';
 import { Capabilities } from '../../domain/capabilities';
 import { ItemAlreadyExists } from '../../domain/errors';
 
 describe('processSubscriptionRequest', () => {
-  it('should insert the first version of subscription-history and subscription', async () => {
+  it('should insert the first version of subscription-history, subscription and activation', async () => {
     const mockEnv = makeTestEnv();
     const testEnv = mockEnv as unknown as Capabilities;
 
+    mockEnv.monotonicIdFn.mockReturnValueOnce({
+      value: anActivationRequest.id,
+    });
     mockEnv.clock.now.mockReturnValueOnce(TE.right(aSubscription.createdAt));
-    mockEnv.hashFn.mockReturnValueOnce(TE.right({ value: aSubscription.id }));
-    mockEnv.hashFn.mockReturnValueOnce(
-      TE.right({ value: aSubscriptionHistory.id }),
-    );
+    mockEnv.hashFn
+      .mockReturnValueOnce(TE.right({ value: aSubscription.id }))
+      .mockReturnValueOnce(TE.right({ value: aSubscriptionHistory.id }));
     mockEnv.subscriptionWriter.insert.mockReturnValueOnce(
       TE.right(aSubscription),
     );
     mockEnv.subscriptionHistoryWriter.insert.mockReturnValueOnce(
       TE.right(aSubscriptionHistory),
+    );
+    mockEnv.activationRequestRepository.insert.mockReturnValueOnce(
+      TE.right(anActivationRequest),
     );
 
     const actual =
@@ -35,22 +41,28 @@ describe('processSubscriptionRequest', () => {
     expect(actual).toStrictEqual(expected);
     expect(mockEnv.subscriptionWriter.insert).toBeCalledTimes(1);
     expect(mockEnv.subscriptionHistoryWriter.insert).toBeCalledTimes(1);
+    expect(mockEnv.activationRequestRepository.insert).toBeCalledTimes(1);
   });
   it('should insert the first version of subscription-history if subscription already exists', async () => {
     const mockEnv = makeTestEnv();
     const testEnv = mockEnv as unknown as Capabilities;
 
+    mockEnv.monotonicIdFn.mockReturnValueOnce({
+      value: anActivationRequest.id,
+    });
     mockEnv.clock.now.mockReturnValueOnce(TE.right(aSubscription.createdAt));
-    mockEnv.hashFn.mockReturnValueOnce(TE.right({ value: aSubscription.id }));
-    mockEnv.hashFn.mockReturnValueOnce(
-      TE.right({ value: aSubscriptionHistory.id }),
-    );
+    mockEnv.hashFn
+      .mockReturnValueOnce(TE.right({ value: aSubscription.id }))
+      .mockReturnValueOnce(TE.right({ value: aSubscriptionHistory.id }));
     mockEnv.subscriptionWriter.insert.mockReturnValueOnce(
       TE.left(new ItemAlreadyExists('')),
     );
     mockEnv.subscriptionHistoryWriter.insert.mockReturnValueOnce(
       TE.right(aSubscriptionHistory),
     );
+    mockEnv.activationRequestRepository.insert.mockReturnValueOnce(
+      TE.right(anActivationRequest),
+    );
 
     const actual =
       await processSubscriptionRequest(aSubscriptionRequest)(testEnv)();
@@ -59,22 +71,28 @@ describe('processSubscriptionRequest', () => {
     expect(actual).toStrictEqual(expected);
     expect(mockEnv.subscriptionWriter.insert).toBeCalledTimes(1);
     expect(mockEnv.subscriptionHistoryWriter.insert).toBeCalledTimes(1);
+    expect(mockEnv.activationRequestRepository.insert).toBeCalledTimes(1);
   });
   it('should insert the subscription if first version of subscription-history already exists', async () => {
     const mockEnv = makeTestEnv();
     const testEnv = mockEnv as unknown as Capabilities;
 
+    mockEnv.monotonicIdFn.mockReturnValueOnce({
+      value: anActivationRequest.id,
+    });
     mockEnv.clock.now.mockReturnValueOnce(TE.right(aSubscription.createdAt));
-    mockEnv.hashFn.mockReturnValueOnce(TE.right({ value: aSubscription.id }));
-    mockEnv.hashFn.mockReturnValueOnce(
-      TE.right({ value: aSubscriptionHistory.id }),
-    );
+    mockEnv.hashFn
+      .mockReturnValueOnce(TE.right({ value: aSubscription.id }))
+      .mockReturnValueOnce(TE.right({ value: aSubscriptionHistory.id }));
     mockEnv.subscriptionWriter.insert.mockReturnValueOnce(
       TE.right(aSubscription),
     );
     mockEnv.subscriptionHistoryWriter.insert.mockReturnValueOnce(
       TE.left(new ItemAlreadyExists('')),
     );
+    mockEnv.activationRequestRepository.insert.mockReturnValueOnce(
+      TE.right(anActivationRequest),
+    );
 
     const actual =
       await processSubscriptionRequest(aSubscriptionRequest)(testEnv)();
@@ -83,22 +101,28 @@ describe('processSubscriptionRequest', () => {
     expect(actual).toStrictEqual(expected);
     expect(mockEnv.subscriptionWriter.insert).toBeCalledTimes(1);
     expect(mockEnv.subscriptionHistoryWriter.insert).toBeCalledTimes(1);
+    expect(mockEnv.activationRequestRepository.insert).toBeCalledTimes(1);
   });
-  it('should succeed if the the first version of subscription-history and subscription already exists', async () => {
+  it('should succeed if the the first version of subscription-history, subscription and activation already exists', async () => {
     const mockEnv = makeTestEnv();
     const testEnv = mockEnv as unknown as Capabilities;
 
+    mockEnv.monotonicIdFn.mockReturnValueOnce({
+      value: anActivationRequest.id,
+    });
     mockEnv.clock.now.mockReturnValueOnce(TE.right(aSubscription.createdAt));
-    mockEnv.hashFn.mockReturnValueOnce(TE.right({ value: aSubscription.id }));
-    mockEnv.hashFn.mockReturnValueOnce(
-      TE.right({ value: aSubscriptionHistory.id }),
-    );
+    mockEnv.hashFn
+      .mockReturnValueOnce(TE.right({ value: aSubscription.id }))
+      .mockReturnValueOnce(TE.right({ value: aSubscriptionHistory.id }));
     mockEnv.subscriptionWriter.insert.mockReturnValueOnce(
       TE.left(new ItemAlreadyExists('')),
     );
     mockEnv.subscriptionHistoryWriter.insert.mockReturnValueOnce(
       TE.left(new ItemAlreadyExists('')),
     );
+    mockEnv.activationRequestRepository.insert.mockReturnValueOnce(
+      TE.left(new ItemAlreadyExists('')),
+    );
 
     const actual =
       await processSubscriptionRequest(aSubscriptionRequest)(testEnv)();
@@ -107,17 +131,20 @@ describe('processSubscriptionRequest', () => {
     expect(actual).toStrictEqual(expected);
     expect(mockEnv.subscriptionWriter.insert).toBeCalledTimes(1);
     expect(mockEnv.subscriptionHistoryWriter.insert).toBeCalledTimes(1);
+    expect(mockEnv.activationRequestRepository.insert).toBeCalledTimes(1);
   });
   it('should fail if an unexpected error is raised', async () => {
     const mockEnv = makeTestEnv();
     const testEnv = mockEnv as unknown as Capabilities;
     const unexpectedError = new Error('Unexpected error');
 
+    mockEnv.monotonicIdFn.mockReturnValueOnce({
+      value: anActivationRequest.id,
+    });
     mockEnv.clock.now.mockReturnValueOnce(TE.right(aSubscription.createdAt));
-    mockEnv.hashFn.mockReturnValueOnce(TE.right({ value: aSubscription.id }));
-    mockEnv.hashFn.mockReturnValueOnce(
-      TE.right({ value: aSubscriptionHistory.id }),
-    );
+    mockEnv.hashFn
+      .mockReturnValueOnce(TE.right({ value: aSubscription.id }))
+      .mockReturnValueOnce(TE.right({ value: aSubscriptionHistory.id }));
     mockEnv.subscriptionWriter.insert.mockReturnValueOnce(
       TE.right(aSubscription),
     );
@@ -132,5 +159,6 @@ describe('processSubscriptionRequest', () => {
     expect(actual).toStrictEqual(expected);
     expect(mockEnv.subscriptionWriter.insert).toBeCalledTimes(1);
     expect(mockEnv.subscriptionHistoryWriter.insert).toBeCalledTimes(1);
+    expect(mockEnv.activationRequestRepository.insert).not.toHaveBeenCalled();
   });
 });